move more platform specific files
