+++ Code/Engine/Foundation/Platform/Android/Basics_Platform.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#if EZ_DISABLED(EZ_PLATFORM_LINUX) && EZ_DISABLED(EZ_PLATFORM_ANDROID)
-#  error "This header should only be included on Linux"
+#if EZ_DISABLED(EZ_PLATFORM_ANDROID)
+#  error "This header should only be included on Android"
 #endif
 
 #include <cstdio>
