+++ Code/Engine/Foundation/Platform/Linux/Screen_Linux.cpp
@@ -2,10 +2,6 @@
 
 #if EZ_ENABLED(EZ_PLATFORM_LINUX)
 
-#  if EZ_ENABLED(EZ_SUPPORTS_GLFW)
-#    include <Foundation/Platform/GLFW/Screen_GLFW.h>
-#  else
-#    include <Foundation/Platform/NoImpl/Screen_NoImpl.h>
-#  endif
+#  include <Foundation/Platform/GLFW/Screen_GLFW.h>
 
 #endif
