+++ Code/Engine/Foundation/Platform/OSX/Screen_OSX.cpp
@@ -2,10 +2,6 @@
 
 #if EZ_ENABLED(EZ_PLATFORM_OSX)
 
-#  if EZ_ENABLED(EZ_SUPPORTS_GLFW)
-#    include <Foundation/Platform/GLFW/Screen_GLFW.h>
-#  else
-#    include <Foundation/Platform/NoImpl/Screen_NoImpl.h>
-#  endif
+#  include <Foundation/Platform/GLFW/Screen_GLFW.h>
 
 #endif
