+++ Code/Tools/HeaderCheck/HeaderCheck.cpp
@@ -407,7 +407,7 @@ public:
       currentFileLower.ToLower();
 
       // ignore includes ending in "_Platform.h", they redirect to platform specific EZ headers
-      const bool ignore = m_IgnoreTarget.m_byName.Contains(includeFileLower) || m_IgnoreSource.m_byName.Contains(currentFileLower) || currentFileLower.EndsWith_NoCase("_Platform.h");
+      const bool ignore = m_IgnoreTarget.m_byName.Contains(includeFileLower) || m_IgnoreSource.m_byName.Contains(currentFileLower) || includeFileLower.EndsWith_NoCase("_Platform.h");
 
       if (!ignore)
       {
